requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
pip==23.3.1
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
setuptools==65.5.0
sqlparse==0.4.4
tzdata==2023.3
